refactor: Move `JsonLayout` conversion to dedicated traits and module.
