+++ src/data/json_model/json_layout.rs
@@ -1,0 +1,41 @@
+use crate::BmaLayout;
+use crate::data::json_model::{JsonLayoutContainer, JsonLayoutVariable};
+use crate::utils::{clone_into_vec, take_if_not_blank};
+use serde::{Deserialize, Serialize};
+
+/// Structure to deserialize JSON info about layout, which contains variables,
+/// containers, and a description.
+///
+/// All of these can be missing in the JSON. If not provided, default empty values
+/// are used.
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub(crate) struct JsonLayout {
+    #[serde(default, rename = "Variables", alias = "variables")]
+    pub variables: Vec<JsonLayoutVariable>,
+    #[serde(default, rename = "Containers", alias = "containers")]
+    pub containers: Vec<JsonLayoutContainer>,
+    #[serde(default, rename = "Description", alias = "description")]
+    pub description: String,
+}
+
+impl From<JsonLayout> for BmaLayout {
+    fn from(value: JsonLayout) -> Self {
+        BmaLayout {
+            variables: clone_into_vec(&value.variables),
+            containers: clone_into_vec(&value.containers),
+            description: take_if_not_blank(value.description.as_str()),
+            zoom_level: None,
+            pan: None,
+        }
+    }
+}
+
+impl From<BmaLayout> for JsonLayout {
+    fn from(value: BmaLayout) -> Self {
+        JsonLayout {
+            variables: clone_into_vec(&value.variables),
+            containers: clone_into_vec(&value.containers),
+            description: value.description.unwrap_or_default(),
+        }
+    }
+}
